First commit
This is a sample of the code used to look for Zprime events where the Zprime decays to a pair of tau leptons, and then one tau decays to an electron+neutrino and the other decays to a muon+neutrino. The bulk of the analysis work is done by EMuAnalysis.C, which is called and fed appropriate configuration parameters by EMuDrv.C. The remainder of the files here are largely for parsing output/making pretty(?) plots.